Move C# API examples to their own projects
