+++ basex-api/src/main/c#/BaseXClient.sln
@@ -1,0 +1,68 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BaseXClient", "BaseXClient\BaseXClient.csproj", "{06A29A09-A45B-49CC-A2D5-DA4E101023D8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AddExample", "AddExample\AddExample.csproj", "{26CFC77A-2DE0-4190-8B06-C28EBAB96071}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution Files", "{6A862760-9D5B-430A-A43F-BA6B1FD17581}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		Directory.Build.props = Directory.Build.props
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateExample", "CreateExample\CreateExample.csproj", "{B6392E5B-84AB-462B-98D5-8F085325B725}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueryBindExample", "QueryBindExample\QueryBindExample.csproj", "{B8C06793-1620-49BA-A443-A8BA96BCFDB8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QueryExample", "QueryExample\QueryExample.csproj", "{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Example", "Example\Example.csproj", "{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{582E3518-7455-452A-85CE-6877043D88A1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventExample", "EventExample\EventExample.csproj", "{E69CE5D9-0825-4981-97C3-0A1F5151E578}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{06A29A09-A45B-49CC-A2D5-DA4E101023D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06A29A09-A45B-49CC-A2D5-DA4E101023D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06A29A09-A45B-49CC-A2D5-DA4E101023D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06A29A09-A45B-49CC-A2D5-DA4E101023D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26CFC77A-2DE0-4190-8B06-C28EBAB96071}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26CFC77A-2DE0-4190-8B06-C28EBAB96071}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26CFC77A-2DE0-4190-8B06-C28EBAB96071}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26CFC77A-2DE0-4190-8B06-C28EBAB96071}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B6392E5B-84AB-462B-98D5-8F085325B725}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B6392E5B-84AB-462B-98D5-8F085325B725}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B6392E5B-84AB-462B-98D5-8F085325B725}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B6392E5B-84AB-462B-98D5-8F085325B725}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8C06793-1620-49BA-A443-A8BA96BCFDB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8C06793-1620-49BA-A443-A8BA96BCFDB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8C06793-1620-49BA-A443-A8BA96BCFDB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8C06793-1620-49BA-A443-A8BA96BCFDB8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E69CE5D9-0825-4981-97C3-0A1F5151E578}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E69CE5D9-0825-4981-97C3-0A1F5151E578}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E69CE5D9-0825-4981-97C3-0A1F5151E578}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E69CE5D9-0825-4981-97C3-0A1F5151E578}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{26CFC77A-2DE0-4190-8B06-C28EBAB96071} = {582E3518-7455-452A-85CE-6877043D88A1}
+		{B6392E5B-84AB-462B-98D5-8F085325B725} = {582E3518-7455-452A-85CE-6877043D88A1}
+		{BB3A9A24-E0D4-444B-BDFD-7EB45C428B3E} = {582E3518-7455-452A-85CE-6877043D88A1}
+		{B8C06793-1620-49BA-A443-A8BA96BCFDB8} = {582E3518-7455-452A-85CE-6877043D88A1}
+		{E7B2DD88-8410-4E71-9C3A-E7CF25D7751A} = {582E3518-7455-452A-85CE-6877043D88A1}
+		{E69CE5D9-0825-4981-97C3-0A1F5151E578} = {582E3518-7455-452A-85CE-6877043D88A1}
+	EndGlobalSection
+EndGlobal
