+++ basex-api/src/main/c#/BaseXClient/BaseXClient.cs
