+++ basex-api/src/main/c#/CreateExample/CreateExample.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BaseXClient\BaseXClient.csproj" />
+  </ItemGroup>
+
+</Project>
