+++ basex-api/src/main/c#/QueryBindExample/QueryBindExample.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BaseXClient\BaseXClient.csproj" />
+  </ItemGroup>
+</Project>
